% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlopeExplorer-package.R
\docType{package}
\name{SlopeExplorer-package}
\alias{SlopeExplorer}
\alias{SlopeExplorer-package}
\title{SlopeExplorer: An Amazing Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: firstname lastname \email{your@email.com}

}
\keyword{internal}
